pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.uv]
index-url = "https://pypi.org/simple"

[project]
name = "flops"
version = "0.1.0"
requires-python = ">=3.13"
dependencies = [
    "anthropic",
    "openai",
    "prompt_toolkit",
    "pyyaml",
    "httpx",
    "rich",
    "bashlex",
    "pydantic",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "autopep8>=2.3.2",
    "autoflake>=2.3.3",
    "vulture>=2.16",
]

[project.scripts]
flops = "flops.cli:main"
[tool.hatch.build.targets.wheel]
packages = ["src/flops"]

[tool.black]
line-length = 100
target-version = ["py313"]
